removing the test of MarkDuplicatesGATK which is out of date
